[WFLY-10288] Remove javax.mail 1.5.6.Final and move to 1.6.1.Final for Java EE 8
