+++ feature-pack/src/main/resources/content/docs/licenses/full-feature-pack-licenses-addendum.html
@@ -22,7 +22,6 @@
 
 <!--
 com.sun.faces:jsf-impl:2.2.13.SP5
-com.sun.mail:javax.mail:1.5.6
 org.apache.lucene:lucene-analyzers-common:5.5.5
 org.apache.lucene:lucene-backward-codecs:5.5.5
 org.apache.lucene:lucene-core:5.5.5
@@ -55,17 +54,6 @@ org.hibernate.javax.persistence:hibernate-jpa-2.1-api:1.0.2.Final
 
     <tr>
         <td>com.sun.faces</td><td>jsf-impl</td><td>2.2.13.SP5</td><td><a href="https://javaee.github.io/glassfish/LICENSE">Common Development and Distribution License 1.1</a>
-        <br>
-        <a href="http://openjdk.java.net/legal/gplv2+ce.html">GNU General Public License v2.0 only, with Classpath exception</a>
-        <br>
-    </td><td><a href="common development and distribution license 1.1.txt">common development and distribution license 1.1.txt</a>
-        <br>
-        <a href="gnu general public license v2.0 only, with classpath exception.txt">gnu general public license v2.0 only, with classpath exception.txt</a>
-        <br>
-    </td>
-    </tr>
-    <tr>
-        <td>com.sun.mail</td><td>javax.mail</td><td>1.5.6</td><td><a href="https://javaee.github.io/glassfish/LICENSE">Common Development and Distribution License 1.1</a>
         <br>
         <a href="http://openjdk.java.net/legal/gplv2+ce.html">GNU General Public License v2.0 only, with Classpath exception</a>
         <br>
