+++ feature-pack/src/main/resources/modules/system/layers/base/javax/mail/api/main/module.xml
@@ -29,15 +29,6 @@
     </dependencies>
 
     <resources>
-        <artifact name="com.sun.mail:javax.mail:1.5.6"> <!-- ${version.javax.mail-1.5} -->
-            <conditions>
-                <property-not-equal name="ee8.preview.mode" value="true"/>
-            </conditions>
-        </artifact>
-        <artifact name="${com.sun.mail:javax.mail}">
-            <conditions>
-                <property-equal name="ee8.preview.mode" value="true"/>
-            </conditions>
-        </artifact>
+        <artifact name="${com.sun.mail:javax.mail}"/>
     </resources>
 </module>
